refactor: вынесение BasicPlayCardList из директории компонента BasicPlayCard
