+++ src/components/section/section.stories.tsx
@@ -1,9 +1,9 @@
 import { ComponentStory, ComponentMeta } from '@storybook/react';
 
-import { Section } from './section';
-import { BasicPlayCardList } from 'components/ui/basic-play-card';
+import { Section } from 'components/section';
+import { BasicPlayCardList } from 'components/ui/basic-play-card-list';
 
-import { Default as DefaultBasicPlayCardList } from 'components/ui/basic-play-card/list/basic-play-card-list.stories';
+import { Default as DefaultBasicPlayCardList } from 'components/ui/basic-play-card-list/basic-play-card-list.stories';
 import { Default as DefaultPersonCardList } from 'components/ui/person-card/list/person-card-list.stories';
 
 export default {
