+++ src/components/ui/basic-play-card-list/basic-play-card-list.module.css
@@ -1,4 +1,4 @@
-.basicPlayCardList {
+.list {
   @mixin hideScrollbar;
 
   --padding: scale(24px);
