+++ src/components/ui/basic-play-card-list/basic-play-card-list.tsx
@@ -2,6 +2,7 @@ import { ReactNode } from 'react';
 import classNames from 'classnames/bind';
 
 import styles from './basic-play-card-list.module.css';
+
 const cx = classNames.bind(styles);
 
 interface IBasicPlayCardListProps extends React.HTMLAttributes<HTMLDivElement> {
@@ -12,7 +13,7 @@ export const BasicPlayCardList = (props: IBasicPlayCardListProps): JSX.Element =
   const { children } = props;
 
   return (
-    <div className={cx('basicPlayCardList')}>
+    <div className={cx('list')}>
       {children}
     </div>
   );
