+++ src/components/ui/basic-play-card/index.ts
@@ -1,2 +1,1 @@
 export * from './basic-play-card';
-export * from './list';
